Move SMTP setting to EasyFramework
