+++ src/ZKEACMS/Controllers/SmtpSettingController.cs
@@ -2,6 +2,7 @@
  * Copyright (c) ZKEASOFT. All rights reserved. 
  * http://www.zkea.net/licenses */
 using Easy.Mvc.Authorize;
+using Easy.Notification;
 using Microsoft.AspNetCore.Mvc;
 using Newtonsoft.Json;
 using System;
